feat: new color scheme; refactor: line class, header files.
