+++ src/PxSitl/include/PxSitl/vision/StateTracking.hpp
@@ -25,6 +25,7 @@
 #include "utils/thresholdtype.hpp"
 #include "utils/RvizPainterObject.hpp"
 #include "utils/RvizPainter.hpp"
+#include "../utils/Line.hpp"
 
 using geometry_msgs::Pose;
 using geometry_msgs::PosePtr;
@@ -35,97 +36,6 @@ using sensor_msgs::CameraInfoConstPtr;
 using visualization_msgs::Marker;
 using visualization_msgs::MarkerConstPtr;
 
-class Line {
-private:
-  tf2::Vector3 _point1;
-  tf2::Vector3 _point2;
-  tf2::Vector3 _midpoint;
-
-public:
-  Line(const tf2::Vector3 &point3d1, const tf2::Vector3 &point3d2) : _point1(point3d1), _point2(point3d2) {
-    _midpoint = tf2::lerp(_point1, _point2, .5);
-  }
-  ~Line() {}
-
-  bool operator==(Line &line) {
-    return _point1 == line._point1 && _point2 == line._point2;
-  }
-
-  inline const tf2::Vector3 &getP1() {
-    return _point1;
-  }
-
-  inline const tf2::Vector3 &getP2() {
-    return _point2;
-  }
-
-  inline const tf2::Vector3 &getMidpoint() {
-    return _midpoint;
-  }
-
-  inline float length2() {
-    return tf2::tf2Distance2(_point2, _point1);
-  }
-
-  inline float length() {
-    return tf2::tf2Distance(_point2, _point1);
-  }
-
-  // S = ah
-  float distToPoint2(const tf2::Vector3 &point) {
-    float dist2ToP1 = tf2::tf2Distance2(point, _point1);
-    float dist2ToP2 = tf2::tf2Distance2(point, _point2);
-
-    auto nearPoint = std::ref(_point1);
-    if (dist2ToP2 < dist2ToP1)
-      nearPoint = std::ref(_point2);
-
-    debug();
-    // ROS_DEBUG("\nDistToP1 %f\nDistToP2 %f", dist2ToP1, dist2ToP2);
-    tf2::Vector3 lineDir = _point2 - _point1;
-    ROS_DEBUG("Line length2 %f", lineDir.length2());
-    if (lineDir.length2() == 0.f)
-      return -1.f;
-
-    tf2::Vector3 numerator = tf2::tf2Cross(point - nearPoint, lineDir); // vector length equals area of parallelogram
-    // ROS_DEBUG("\nNum %f\nDiv %f", numerator.length2(), lineDir.length2());
-
-    return numerator.length2() / lineDir.length2();
-  }
-
-  float distToPoint(tf2::Vector3 &point) {
-    float res = distToPoint2(point);
-    return res == -1.f ? -1.f : sqrtf(res);
-  }
-
-  /**
-   * https://gamedev.stackexchange.com/questions/72528/how-can-i-project-a-3d-point-onto-a-3d-line
-   *
-   * A + dot(AP,AB) / dot(AB,AB) * AB
-   * A = line point1
-   * B = line point2
-   * P = incoming point
-   */
-  tf2::Vector3 porjectPoint(const tf2::Vector3 &point) {
-    tf2::Vector3 AB = _point2 - _point1;
-    tf2::Vector3 AP = point - _point1;
-
-    tf2::Vector3 res = _point1;
-
-    float numerator = tf2::tf2Dot(AP, AB);
-    float denominator = tf2::tf2Dot(AB, AB);
-
-    res = res + numerator / denominator * AB;
-
-    return res;
-  }
-
-  void debug() {
-    ROS_DEBUG("P1 %f %f %f", _point1.x(), _point1.y(), _point1.z());
-    ROS_DEBUG("P2 %f %f %f", _point2.x(), _point2.y(), _point2.z());
-  }
-};
-
 class StateTracking : public State {
 private:
   int _fps = 0;
@@ -135,8 +45,6 @@ private:
   const char *_winName = "Tracking";
   float _prevDist = 0.f;
   float _dt4prediction = 0.01f;
-
-  int tmpMarkerIndex = 1;
 
   BallTracking *_bt = nullptr;
   RosVH *_vh = nullptr;
@@ -152,9 +60,8 @@ private:
   ros::Time _startTrackingTimer;
   ros::Time _resetTimer;
   ros::Time _detectionTimer;
-  ros::Time _buildRealTrekLineTimer;
   ros::Time _aTimer;
-  // ros::Publisher _ballPub;
+
   tf2_ros::Buffer _tfBuffer;
   tf2_ros::TransformListener *_tfListener;
   tf2::Vector3 _firstObjPosition;
@@ -163,29 +70,16 @@ private:
   CameraInfoConstPtr _cameraInfo;
   PinholeCameraModel _cameraModel;
 
-  RvizPainter *_rvizPainter;
+  RvizPainter *_rvizPainter; //< TODO: smart pointer
   RvizPainterObject _rvizPainterObject;
 
   cv::Mat _frame;
   cv::Mat _depth;
 
-  /* void drawLine(const tf2::Vector3 &p1, const tf2::Vector3 &p2, const std_msgs::ColorRGBA &c);
-  void drawLine(geometry_msgs::Point &p1, geometry_msgs::Point &p2, const std_msgs::ColorRGBA &c);
-  void drawArrow(const Pose &pose, const std_msgs::ColorRGBA &c, const char *name);
-  void drawArrow(const tf2::Vector3 &position, const tf2::Quaternion &orientation, const std_msgs::ColorRGBA &c, const char *name);
-  void drawObjPose(Pose &p);
-  void drawObjPose(Pose &p, const std_msgs::ColorRGBA &c);
-  void drawObjPose(geometry_msgs::Point &p, const std_msgs::ColorRGBA &c);
-  void drawObjPose(const tf2::Vector3 &position, const std_msgs::ColorRGBA &c);
-  void drawObjRealLine(std::list<geometry_msgs::Point> &list);
-  void drawObjRealLine(const tf2::Vector3 &position, const std_msgs::ColorRGBA &c);
-  void drawObjPredictedLine(std::list<geometry_msgs::Point> &list); */
-  // void pubMarker(Marker m);
   void transformPose(tf2::Vector3 &position);
   Pose transformPose2(tf2::Vector3 &position, const tf2::Quaternion &orientation);
   void conceptOne(cv::Mat &mask, cv::Point2i &center, uint16_t &radius);
   void conceptTwo(cv::Mat &mask, cv::Point2i &point2d, uint16_t &radius);
-  float getDistToObj(cv::Mat &mask, uint16_t &radius);
 
   /**
    * Builds a 3D object position from 2D Image via camera model
@@ -195,6 +89,10 @@ private:
    * @param[out] objPos tf2::Vector3 - 3d point
    **/
   void getObjPosFromImg(cv::Point2i &point2d, float distToObj, tf2::Vector3 &objPos);
+  
+  float getDistToObj(cv::Mat &mask, uint16_t &radius);
+  float getVelocity(float x, float y, float angle);
+  float getContactProbobility(tf2::Vector3 &currentPosition, tf2::Vector3 &cameraPosition);
 
 public:
   StateTracking(BallTrackingRos &context, ros::NodeHandle &nh) : State(context, "Tracking"), _nh(nh) {}
@@ -206,89 +104,6 @@ public:
   void tracking() override {}
   void wait() override;
   void execute() override;
-
-  float getVelocity(float x, float y, float angle) {
-    float num = PZ_GRAVITY * powf(x, 2);
-    float den = 2.0f * (y - tanf(angle) * x) * powf(cosf(angle), 2);
-    float v2 = num / den;
-    return sqrtf(abs(v2));
-  }
-
-  float getContactProbobility(tf2::Vector3 &currentPosition, tf2::Vector3 &cameraPosition) {
-    if (_predictedSigments.size() == 0)
-      return -1;
-
-    uint16_t realTrekLen = _realTraj.size();
-    uint16_t predictedTrekLen = _predTrajectory.size();
-    uint16_t realTrekMiddle = realTrekLen / 2;
-    uint16_t predictedTrekMiddle = predictedTrekLen / 2;
-
-    // ROS_DEBUG("RealTrekLine size: %i\nMiddle: %i", _realTraj.size(), middle);
-    // geometry_msgs::Point curRealPoint;
-    uint16_t realTrekPosCounter = 0;
-    static uint8_t colorCounter = 0;
-
-    float minDist = 100.0f;
-    auto nearLine = std::ref(_predictedSigments[0]);
-    auto middle = std::cref(nearLine.get().getMidpoint());
-
-    for (auto &&line : _predictedSigments) {
-      const tf2::Vector3 &tmpMiddle = line.getMidpoint();
-      float tmpDist = tf2::tf2Distance2(tmpMiddle, currentPosition);
-
-      if (tmpDist < minDist) {
-        nearLine = std::ref(line);
-        middle = std::cref(tmpMiddle);
-        minDist = tmpDist;
-      }
-    }
-
-    float distToPoint = nearLine.get().distToPoint2(currentPosition);
-
-    if (distToPoint == -1.f)
-      return -1.f;
-
-    // drawLine(currentPosition, middle, Utils::Colors.at(Utils::Color::Black));
-    tf2::Vector3 pointOnLine = nearLine.get().porjectPoint(currentPosition);
-
-    // drawObjPose(pointOnLine, Utils::Colors.at(Utils::Color::AlfaBlue));
-
-    tf2::Vector3 fromPointToLine = pointOnLine - currentPosition;
-
-    float euclidian = fromPointToLine.length();
-    float chebyshev = std::max(std::max(fromPointToLine.x(), fromPointToLine.y()), fromPointToLine.z());
-
-    float c = (currentPosition - cameraPosition).length2();
-    float aCath = distToPoint;
-    float sinMetric = aCath / c;
-    float bCath = (cameraPosition - pointOnLine).length2();
-    float cosMetric = bCath / c;
-
-    float totalDistToObject = (cameraPosition - _firstObjPosition).length2();
-    float passDist = (_firstObjPosition - pointOnLine).length2();
-    float passDistMetric = passDist / totalDistToObject;
-
-    /* if (_prevDist != 0) {
-      float deltaDist = euclidian - _prevDist;
-      double deltaTime = (ros::Time::now() - _aTimer).toSec();
-      ROS_INFO("a %lf", deltaDist / deltaTime);
-    }
-
-    _prevDist = euclidian;
-    _aTimer = ros::Time::now(); */
-
-    float avrMetric = (cosMetric + passDistMetric) / 2.f;
-
-    ROS_DEBUG_NAMED("dist_to_line", "DistToLine: %f", sqrtf(distToPoint));
-    ROS_DEBUG_NAMED("euclidian", "Euclidian %f", euclidian);
-    ROS_DEBUG_NAMED("chebyshev", "Chebyshev %f", chebyshev);
-    ROS_DEBUG_NAMED("sin", "Sin %f", 1.f - sinMetric);
-    ROS_DEBUG_NAMED("cos", "Cos %f", cosMetric);
-    ROS_DEBUG_NAMED("pass_dist", "PossDist %f", passDistMetric);
-    ROS_DEBUG_NAMED("avr", "Avr %f", avrMetric);
-
-    return 1;
-  }
 };
 
 #endif // _VISION_STATE_TRACKING_H_
